+++ demisto_sdk/commands/format/update_generic_json.py
@@ -26,7 +26,7 @@ class BaseUpdateJSON(BaseUpdate):
                  verbose: bool = False,
                  **kwargs):
         super().__init__(input=input, output=output, path=path, from_version=from_version, no_validate=no_validate,
-                         verbose=verbose)
+                         verbose=verbose, **kwargs)
 
     def set_default_values_as_needed(self):
         """Sets basic arguments of reputation commands to be default, isArray and required."""