+++ demisto_sdk/tests/test_files/validate_integration_test_valid_types.py
@@ -477,7 +477,7 @@ REPUTATION = {
     "defaultMapping": {},
     "manualMapping": None,
     "fileHashesPriority": None,
-    "legacyName": ["Malware"]
+    "legacyNames": ["Malware"]
 }
 
 WIDGET = {